:recycle: Refector : Change Folder structure and simplifiy layout file
