+++ app/(sign)/layout.tsx
@@ -1,5 +1,4 @@
 import Head from 'next/head';
-import '../../styles/globals.css';
 
 export const metadata = {
   title: 'Next.js',
