+++ app/layout.tsx
@@ -1,3 +1,4 @@
+import { PropsWithChildren } from 'react';
 import '../styles/globals.css';
 
 export const metadata = {
@@ -5,9 +6,9 @@ export const metadata = {
   description: 'Generated by Next.js',
 };
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+export default function RootLayout({ children }: PropsWithChildren) {
   return (
-    <html lang='en'>
+    <html lang='ko'>
       <body>{children}</body>
     </html>
   );
