--- app/signin/layout.tsx
@@ -1,27 +1,0 @@
-import Head from 'next/head';
-import '../../styles/globals.css';
-
-export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js',
-};
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <html lang='ko'>
-      <Head>
-        <title>Taskify</title>
-        <meta name='viewport' content='minimum-scale=1, initial-scale=1, width=device-width' />
-        <meta
-          name='description'
-          content={
-            '친구와 일정을 만들고 직접 관리해보세요! 꼭 필요한 기능만이 들어가 속도가 빠르고 모바일과 PC 환경 모두 사용할 수 있어요😀'
-          }
-        />
-        <meta property='og:title' content={'Taskify: 일정 관리 서비스'} />
-        <meta property='og:type' content='website' />
-      </Head>
-      <body className='w-full h-screen'>{children}</body>
-    </html>
-  );
-}
